mise/requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23ubuntu1
entrypoints==0.3
future==0.18.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
MarkupSafe==1.1.0
more-itertools==4.2.0
mysql-connector-python==8.0.32
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pip==20.0.2
protobuf==3.20.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36ubuntu1
python-docx==0.8.11
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
sos==4.4
ssh-import-id==5.1
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
webexteamssdk==1.6.1
wheel==0.34.2
zipp==1.0.0
zope.interface==4.7.1

